moved stackAnnsByType to utils.
